+++ src/data/vector.cpp
@@ -13,7 +13,7 @@ extern BError* INCOMPATIBLE_SIZES;
 
 #define POS(v, i, j) (i * v + j)
 
-Vector::Vector(int size) : data(new double[size]), size(size), lockable(0), Data(VECTOR) {
+Vector::Vector(uint64_t size) : data(new double[size]), size(size), lockable(0), Data(VECTOR) {
     dims = new int[1];
     dims[0] = size;
     ndims = 1;
@@ -21,13 +21,13 @@ Vector::Vector(int size) : data(new double[size]), size(size), lockable(0), Data
     natdims = 0;
 }
 
-Vector::Vector(int size, bool setToZero) : Vector(size) {
+Vector::Vector(uint64_t size, bool setToZero) : Vector(size) {
     if (setToZero) {
         std::fill(data, data + size, 0);
     }
 }
 
-Vector::Vector(int size1, int size2) : data(new double[size1 * size2]), size(size1 * size2), lockable(0), Data(VECTOR) {
+Vector::Vector(uint64_t size1, uint64_t size2) : data(new double[size1 * size2]), size(size1 * size2), lockable(0), Data(VECTOR) {
     dims = new int[2];
     dims[0] = size1;
     dims[1] = size2;
@@ -36,7 +36,7 @@ Vector::Vector(int size1, int size2) : data(new double[size1 * size2]), size(siz
     natdims = 0;
 }
 
-Vector::Vector(int size1, int size2, bool setToZero) : Vector(size1, size2) {
+Vector::Vector(uint64_t size1, uint64_t size2, bool setToZero) : Vector(size1, size2) {
     if (setToZero) {
         std::fill(data, data + size, 0);
     }
@@ -50,7 +50,7 @@ Vector::Vector(double* data, const Vector* prototype) : data(data), size(prototy
     natdims = prototype->natdims;
 }
 
-Vector::Vector(double* data, const Vector* prototype, int new_dim_access) : data(data), size(prototype->size), lockable(0), Data(VECTOR) {
+Vector::Vector(double* data, const Vector* prototype, uint64_t new_dim_access) : data(data), size(prototype->size), lockable(0), Data(VECTOR) {
     ndims = prototype->ndims;
     dims = new int[ndims];
     std::copy(prototype->dims, prototype->dims + ndims, dims);
@@ -60,7 +60,7 @@ Vector::Vector(double* data, const Vector* prototype, int new_dim_access) : data
     natdims = prototype->natdims + 1;
 }
 
-Vector::Vector(double* data, int size1, int size2) : data(data), size(size1 * size2), lockable(0), Data(VECTOR) {
+Vector::Vector(double* data, uint64_t size1, uint64_t size2) : data(data), size(size1 * size2), lockable(0), Data(VECTOR) {
     dims = new int[2];
     dims[0] = size1;
     dims[1] = size2;
